packages/zec-rpc-gateway/requirements.txt
apifairy==1.3.0
apispec==6.3.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
astroid==3.0.0
Authlib==1.2.1
blinker==1.6.2
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
click==8.1.7
coverage==7.3.2
cryptography==41.0.4
dill==0.3.7
docker==6.1.3
exceptiongroup==1.1.3
Flask==3.0.0
Flask-Cors==4.0.0
Flask-HTTPAuth==4.8.0
Flask-Login==0.6.2
flask-marshmallow==0.15.0
Flask-Script==2.0.6
gunicorn==21.2.0
idna==3.4
importlib-metadata==6.8.0
iniconfig==2.0.0
isort==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonrpcclient==4.0.3
lazy-object-proxy==1.9.0
MarkupSafe==2.1.3
marshmallow==3.20.1
mccabe==0.7.0
nanoid==2.0.0
packaging==23.2
passlib==1.7.4
platformdirs==3.11.0
pluggy==1.3.0
pycparser==2.21
pylint==3.0.0
pytest==7.4.2
pytest-cov==4.1.0
python-dotenv==1.0.0
requests==2.31.0
six==1.16.0
tomli==2.0.1
tomlkit==0.12.1
typing_extensions==4.8.0
urllib3==2.0.6
webargs==8.3.0
websocket-client==1.6.3
Werkzeug==3.0.0
wrapt==1.15.0
zecwallet==1.5.1
zipp==3.17.0
